move footerterm to userstack components directory
