+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export { default as ArticleBox } from './ArticleBox';
 export { default as Button } from './Button';
-export { default as FootTerms } from './FootTerms';
 export { default as Header } from './Header';
 export { default as InfoContainer } from './InfoContainer';
 export { default as Profile } from './ProfileBox';
