+++ src/screens/UserStack/components/FootTerms/FootTerms.style.ts
